Make gitignore so that we can see the build deps.
Here's a funny thing. Although the deps are added
as "submodules", they aren't git submodule init'd.
So the parent module doesn't see them as submodules
in github.

Maybe? I'm not sure. At any rate, in vscode I can see
these repositories as editable repositories with remotes that (in many cases) I control. Or forks of which I can make if I
need patches. So, now, as I edit this, I can adjust any of these
repositories which I need to fork or which are mine originally. I encourage you to do the same!